refactor(registry): move tag-resolution and env tag validator to utils
Colocate shared tag helpers with other registry utilities. CI runs the validator from script/registry/utils; abi-registry imports ./utils/tag-resolution.js. README and workflow paths updated.
